LSP: Improve memory usage by reducing clones of file paths and shrinking `CodeLocation` and `ColumnInfo` (#7020)
* Initial work to optimize some memory usage for CodeLocation

* Remove expanded

* Using u32 instead of usize for fields in CodeLocation types

* Added changelog

* Formatting. Renaming CodeLocation to CodeLocationWithFile. Re-using CodeLocation from dbt_common::error

* code quality

* code quality

* Code quality

* code quality

* code quality

* Do not need this function anymore

* do not need expanded_file in release

* Code quality

* code quality

* code quality

* We do need the expanded_file for release

* In LSP, just set it to None

* Use Arc<PathBuf>

* formatting

* removed unused fields for ColumnInfo

* code quality

* code quality

* Feedback

* Feedback and fixing build

* Feedback

GitOrigin-RevId: b42789779d94fc186645b6e88320251755971025

+++ crates/dbt-frontend-common/src/span.rs
@@ -25,7 +25,7 @@ impl Span {
         // input[self.start.index..self.stop.index + 1].to_string()
         // because macro span is exclusive, we removed +1
         // be careful lint fix is inclusive, lint fix will break
-        input[self.start.index..self.stop.index].to_string()
+        input[self.start.index as usize..self.stop.index as usize].to_string()
     }
 
     pub fn with_offset(&self, start_location: &CodeLocation) -> Self {
